#18 - loading msg while login function
