+++ src/components/Footer/Footer.js
@@ -9,14 +9,13 @@ export const Footer = (props) =>{
   return (
 
     <div className="Footer">
-      <Menu icon="labeled" inverted fixed="bottom" widths="4">
+      <Menu icon="labeled" inverted fixed="bottom" widths="5">
         <Link to="/Home">
 
           <Menu.Item >
             <Icon name="h" className={props.active==="/Home" ? "active":""}/>
-            Home
+            <span id="labelHome">Home</span>
           </Menu.Item>
-
         </Link>
 
         <Link to="/Games">
@@ -35,7 +34,7 @@ export const Footer = (props) =>{
         <Link to="/Profile">
           <Menu.Item>
             <Icon name="user" className={props.active==="/Profile" ? "active":""} />
-            {props.signedIn? "Signed In": "not Signed in"}
+            {props.user ? "signed in": "no user"}
           </Menu.Item>
         </Link>
 
@@ -51,29 +50,3 @@ export const Footer = (props) =>{
 
   )
 }
-
-
-/*
-<Link to="/Home">Home</Link>
-<Link to="/Games">Games</Link>
-<Link to="/Score">Score</Link>
-
-<Menu icon="labeled" inverted fixed="bottom" widths="4">
-  <Menu.Item to="/Home" >
-    <Icon name="h" />
-    Home
-  </Menu.Item>
-  <Menu.Item to="/Games">
-    <Icon name="gamepad" />
-    Games
-  </Menu.Item>
-  <Menu.Item to="/Table">
-    <Icon name="table" />
-    Scores
-  </Menu.Item>
-  <Menu.Item to="/Info">
-    <Icon name="info" />
-    Info
-  </Menu.Item>
-</Menu>
-*/
