+++ src/components/scoreboard/scoreboard.css
@@ -13,20 +13,20 @@
     background-color: black;
 }*/
 
-.teamAway, .teamHome {
+#scoreboard >.teamAway, #scoreboard >.teamHome {
     height: auto;
     background-color: #313841;
     width: 100%;
 }
 
-.score {
+#scoreboard > .score {
     height: 50px;
     line-height: 50px;
     text-align: center;
     font-size: 2.5em;
 }
 
-.player {
+#scoreboard > .player {
     display: inline-block;
     width: 49%;
     height: 100px;
@@ -36,11 +36,11 @@
     color: #B2CB5B;
 }
 
-.playerName {
+#scoreboard > .playerName {
     margin: 10px;
 }
 
-.player button {
+#scoreboard > .player button {
     width: 25px;
     padding: 7px;
     border-radius: 5px;
@@ -48,20 +48,20 @@
     top: -7px;
 }
 
-.player span {
+#scoreboard > .player span {
     border: 0px solid rgba(236, 236, 236, 0.54);
     padding: 5px;
     margin-left: 10px;
     margin-right: 10px;
 }
 
-.danger {
+#scoreboard > .danger {
     background-color: rgb(194, 41, 0);
 }
-.primary{
+#scoreboard > .primary{
     background-color: #B2CB5B;
 }
 
-.home, .away {
+#scoreboard > .home, #scoreboard > .away {
     position: absolute;
 }
